dom/base/test/test_anonymousContent_canvas.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1212477
-->
<meta charset="utf-8">
<title>Test for Bug 1212477 - Needs a way to access to &lt;canvas&gt;'s context (2d, webgl) from Anonymous Content API</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1212477">Mozilla Bug 1212477</a>
<div>
  <div id="id" class="test">text content</div>
  <canvas id="canvas2d"></canvas>
  <canvas id="canvas-webgl"></canvas>
  <canvas id="canvas-foo"></canvas>
</div>
<script>
  let chromeDocument = SpecialPowers.wrap(document);
  let testElement = document.querySelector("div");

  let anonymousContent = chromeDocument.insertAnonymousContent();
  anonymousContent.root.appendChild(testElement.cloneNode(true));

  let canvas2d = anonymousContent.root.getElementById("canvas2d");
  let context2d = canvas2d.getContext("2d");

  is(context2d.toString(), "[object CanvasRenderingContext2D]",
      "2D Context is returned properly");

  is(context2d.canvas, null, "context's canvas property");

  is(anonymousContent.root.getElementById("canvas-foo").getContext("foo"),
     null,
     "Context is null for unknown context type");

  const normalWebGL = document.createElement('canvas').getContext('webgl');
  if (normalWebGL) {
    let canvasWebGL = anonymousContent.root.getElementById("canvas-webgl");
    let webgl = canvasWebGL.getContext("webgl");

    is(webgl.toString(), "[object WebGLRenderingContext]",
        "WebGL Context is returned properly");

    is(webgl.canvas, null,
        "WebGL context's canvas property is null in anonymous content");
  }
  chromeDocument.removeAnonymousContent(anonymousContent);
</script>
